backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
websockets==13.1

# Computer Vision & ML
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
torch==2.5.1
torchvision==0.20.1

# Async & Performance
aiofiles==24.1.0
python-jose[cryptography]==3.3.0

# Monitoring & Logging
python-json-logger==2.0.7

# Environment & Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
